# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example for response examples value",
        "description": "Example info",
        "contact": {
            "name": "fitrazh@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8070",
            "description": "Local API server"
        }
    ],
    "paths": {
        "/friend/request": {
            "post": {
                "tags": [
                    "Friend"
                ],
                "summary": "Friend request",
                "description": "Send friend request",
                "operationId": "v1FriendRequest",
                "requestBody": {
                    "request": "Friend request",
                    "description": "Required and optional input to update an order product.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "requestor": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "email1@example.com"
                                    },
                                    "to": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "email2@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "requestor": {
                                            "type": "string",
                                            "example": "The selected requestor is invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "429": {
                        "description": "Too many request"
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/friend/accept": {
            "put": {
                "tags": [
                    "Friend"
                ],
                "summary": "Accept friend request",
                "description": "Accept friend request",
                "operationId": "v1UpdateFriendRequestAccept",
                "requestBody": {
                    "request": "Friend request",
                    "description": "Required and optional input to update an order product.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "requestor": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "email1@example.com"
                                    },
                                    "to": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "email2@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "requestor": {
                                            "type": "string",
                                            "example": "The selected requestor is invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "429": {
                        "description": "Too many request"
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/friend/reject": {
            "put": {
                "tags": [
                    "Friend"
                ],
                "summary": "Reject friend request",
                "description": "Reject friend request",
                "operationId": "v1UpdateFriendRequestReject",
                "requestBody": {
                    "request": "Friend request",
                    "description": "Required and optional input to update an order product.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "requestor": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "email1@example.com"
                                    },
                                    "to": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "email2@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "requestor": {
                                            "type": "string",
                                            "example": "The selected requestor is invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "429": {
                        "description": "Too many request"
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/friend/list": {
            "post": {
                "tags": [
                    "Friend"
                ],
                "summary": "API to see list of friends",
                "description": "API to see list of friends",
                "operationId": "v1GetFriendList",
                "requestBody": {
                    "request": "Friend request",
                    "description": "Required and optional input to update an order product.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "email1@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "requestor": {
                                            "type": "string",
                                            "example": "The selected requestor is invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "429": {
                        "description": "Too many request"
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "requestor": {
                                                        "type": "string",
                                                        "example": "email1@mail.com"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "reject"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}